back/src/requirements.txt
Django>=2.0,<3.0
django-environ==0.4.5
djangorestframework==3.10.3
django-cors-headers==3.1
django-heroku==0.3.1
django-storages==1.7.2
django-filter==2.2
gunicorn==19
pillow==6.1
boto3==1.9


